:sparkles: added responsiveness
