+++ src/core/index.ts
@@ -1,0 +1,4 @@
+export { default as scene } from "./scene";
+export { default as camera } from "./camera";
+export { default as renderer } from "./renderer";
+export { default } from "./setup";
