+++ src/core/renderer.ts
@@ -1,0 +1,6 @@
+import { WebGLRenderer } from "three";
+
+const renderer = new WebGLRenderer();
+renderer.setSize(window.innerWidth, window.innerHeight);
+
+export default renderer;
