+++ src/core/scene.ts
@@ -1,0 +1,5 @@
+import { Scene } from "three";
+
+const scene = new Scene();
+
+export default scene;
