+++ src/events/index.ts
@@ -1,0 +1,7 @@
+import initWindowResize from "./onWindowResize";
+
+function initEvents() {
+  initWindowResize();
+}
+
+export default initEvents;
