+++ src/shapes/index.ts
@@ -1,0 +1,1 @@
+export { default as cube } from "./cube";
